Restricted scope of Memory Appender
