+++ utils/src/test/resources/log4j.properties
@@ -12,4 +12,3 @@ log4j.logger.org.eclipse.jetty=ERROR
 
 # TransmogrifAI logging
 log4j.logger.com.salesforce.op=ERROR
-log4j.logger.com.salesforce.op.utils.spark.OpSparkListener=INFO
